// lowcode/glitterBundle/helper/storage.d.ts
export declare const Storage: {
    editor_mode: 'user' | 'dev';
    select_page_type: 'page' | 'article' | 'module' | 'blog' | 'backend';
    lastSelect: string;
    view_type: 'mobile' | 'desktop' | 'col3' | 'fullScreen';
    select_add_btn: 'me' | 'official' | "view" | 'code' | 'ai' | "template" | 'copy';
    select_function: 'page-editor' | 'backend-manger' | 'server-manager' | 'user-editor';
    page_set_select: 'normal' | 'seo' | 'module' | 'code' | 'form';
    develop_mode: 'false' | 'true';
    global_select: "domain" | 'plugin' | 'resource' | 'event' | 'style' | 'code';
    select_global_editor_tab: 'view' | 'event';
    code_set_select: 'style' | 'script' | 'code' | 'form';
    select_bg_btn: 'official' | 'custom';
    select_item: any;
    page_setting_item: 'layout' | 'color' | 'widget';
    page_setting_global: 'editor' | 'setting';
};